# pagos/src_2/src/main/resources/openapi.yml
openapi: 3.0.3
info:
  version: "1.0.0"
  title: API Tramite y Rectificacion Query
  description: API para la consulta  de tramites y rectificaciones
servers:
  - url: http://localhost:8025/cp2/tramiteyrectificacion-query
    description: Generated server url
...
paths:
  /tramites:
    get:
      tags:
        - TRAMITE
      summary: Obtiene el listado de tramites
      description: Permite obtener el listado de tramites
      operationId: obtenerTramites
      parameters:
        - in: query
          name: queryParams
          required: true
          schema:
            $ref: "#/components/schemas/GetTramiteQueryParamsDto"
      responses:
        "200":
          description: Búsqueda Exitosa.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseTramiteDto"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCpGeneric"
        "500":
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCpGeneric"
  /tramites/pago/regla/{escalaId}:
    get:
      tags:
        - TRAMITE
      summary: Obtiene los tramites por escala
      description: Permite obtener los tramites por escala
      operationId: obtenerTramiteReglaPago
      parameters:
        - in: path
          name: escalaId
          required: true
          schema:
            type: integer
          description: Identificador de la escala
        - in: query
          name: documentoId
          required: false
          schema:
            type: integer
          description: Identificador del documento
        - in: query
          name: indicadorEs
          required: false
          schema:
            type: string
          description: Indicador entrada o salida
      responses:
        "200":
          description: Búsqueda Exitosa.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseTramitePagoDto"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCpGeneric"
        "500":
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCpGeneric"
  /tramites/{id}:
    get:
      tags:
        - TRAMITE
      summary: Obtiene el tramite
      description: Permite obtener el tramite
      operationId: obtenerTramite
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Identificador del tramite
      responses:
        "200":
          description: Búsqueda Exitosa.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseTramiteDetalleDto"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCpGeneric"
        "500":
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCpGeneric"
  /tramites/escala/{escalaId}/documento/{documentoId}:
    get:
      tags:
        - TRAMITE
      summary: Obtiene un único trámite basado en escala, documento e indicador
      description: Retorna un único trámite basado en los identificadores de escala, documento y un indicador adicional.
      operationId: obtenerTramitePorEscalaYDocumento
      parameters:
        - in: path
          name: escalaId
          required: true
          schema:
            type: integer
          description: Identificador de la escala del trámite.
        - in: path
          name: documentoId
          required: true
          schema:
            type: integer
          description: Identificador del documento asociado al trámite.
      responses:
        "200":
          description: Trámite encontrado exitosamente.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseTramiteDetalleDto"
        "400":
          description: Solicitud incorrecta.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCpGeneric"
        "404":
          description: Trámite no encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCpGeneric"
        "500":
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCpGeneric"
...
components:
  schemas:
    ApiResponseCpGeneric:
      type: object
      properties:
        meta:
          $ref: "#/components/schemas/ApiResponseMetadata"
        data:
          type: array
          items:
            type: object # This can be overridden by specific schemas
    ApiResponseMetadata:
      type: object
      properties:
        result:
          type: string
        mensajes:
          type: array
          items:
            $ref: "#/components/schemas/MetadataMessage"
          default: []
        cantidadRegistros:
          type: integer
          default: 0
        cantidadRegistrosTotal:
          type: integer
          nullable: true
        atributos:
          type: object
          additionalProperties:
            type: string
          description: Atributos adicionales de la respuesta
    MetadataMessage:
      title: MetadataMessage
      type: object
      properties:
        codigo:
          type: string
          description: Codigo de error de Sistema
        tipo:
          type: string
          description: Tipo de Error de Sistema
        mensaje:
          type: string
          description: Descripcion del error de Sistema
        parametrosDeMensaje:
          type: array
          nullable: true
          items:
            type: string
          description: Parametros del mensaje de error
...
    GetTramiteQueryParamsDto:
      title: GetTramiteQueryParamsDto
      type: object
      required:
        - pageNumber
        - pageSize
      properties:
        tupa:
          type: string
          description: Tupa del tramite
        expedientes:
          type: boolean
          description: Indica si se desea obtener los expedientes
        due:
          type: string
          description: Due del tramite
        nombreNave:
          type: string
          description: Nombre de la nave
        entidad:
          type: integer
          description: Identificador de la entidad
        estadoTramite:
          type: string
          description: Identificador del estado del tramite
        numeroSuce:
          type: string
          description: Numero de suce
        numeroTramite:
          type: string
          description: Numero de tramite
        cpb:
          type: string
          description: Codigo de barras
        agencia:
          type: integer
          description: Identificador de la agencia
        tramiteFechaDesde:
          type: string
          example: "2021-01-01"
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
          description: "Date in the format YYYY-MM-DD"
        tramiteFechaHasta:
          type: string
          example: "2021-01-01"
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
          description: "Date in the format YYYY-MM-DD"
        pageNumber:
          type: integer
          description: Numero de pagina
        pageSize:
          type: integer
          description: Tamaño de pagina
    ApiResponseTramiteDto:
      allOf:
        - $ref: "#/components/schemas/ApiResponseCpGeneric"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/TramiteDto"
    ApiResponseTramitePagoDto:
      allOf:
        - $ref: "#/components/schemas/ApiResponseCpGeneric"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/TramitePagoDto"
    TramitePagoDto:
      type: object
      properties:
        tramiteId:
          type: integer
          description: Identificador del tramite
        escalaId:
          type: integer
          description: Identificador de la escala
        reglaPagoExencionAplicada:
          type: string
          description: Regla de pago o exención aplicada al tramite
    ApiResponseTramiteDetalleDto:
      allOf:
        - $ref: "#/components/schemas/ApiResponseCpGeneric"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/TramiteDetalleDto"
...
    TramiteDto:
      type: object
      properties:
        id:
          type: integer
          description: Identificador del tramite
        due:
          type: string
          description: Due del tramite
        nombreNave:
          type: string
          description: Nombre de la nave
        numeroSuce:
          type: string
          description: Numero de suce
        numeroTramite:
          type: string
          description: Numero de tramite
        entidadNombre:
          type: string
          description: Nombre de la entidad
        tupa:
          type: string
          description: Tupa del tramite
        estadoTramite:
          type: string
          description: Identificador del estado del tramite
        cpb:
          type: string
          description: Codigo de barras
        monto:
          type: number
          description: Monto del tramite
        estadoDePago:
          type: string
          description: Estado de pago del tramite
        agenciaNombre:
          type: string
          description: Nombre de la agencia
        fechaDeclaracionJuradaActual:
          type: string
          format: date-time
          description: Fecha de la declaracion jurada
        declaracionesJuradasCantidad:
          type: integer
          description: Numero de declaraciones juradas
        descripcion:
          type: string
          description: Descripcion del tramite
        reglaPagoExencionAplicada:
          type: string
          description: Regla de pago o exención aplicada al tramite
        documento:
          type: string
          description: Documento del tramite
        indNoRequierePago:
          type: boolean
          description: Indicador de si el tramite no requiere pago
        fechaTramite:
          type: string
          format: date-time
          description: Fecha del tramite
    TramiteDetalleDto:
      type: object
      properties:
        id:
          type: integer
          description: Identificador del tramite
        due:
          type: string
          description: Due del tramite
        nombreNave:
          type: string
          description: Nombre de la nave
        numeroSuce:
          type: string
          description: Numero de suce
        numeroTramite:
          type: string
          description: Numero de tramite
        entidadNombre:
          type: string
          description: Nombre de la entidad
        tupa:
          type: string
          description: Tupa del tramite
        estadoTramite:
          type: string
          description: Identificador del estado del tramite
        cpb:
          type: string
          description: Codigo de barras
        monto:
          type: number
          description: Monto del tramite
        estadoDePago:
          type: string
          description: Estado de pago del tramite
        agenciaNombre:
          type: string
          description: Nombre de la agencia
        fechaDeclaracionJuradaActual:
          type: string
          format: date-time
          description: Fecha de la declaracion jurada
        declaracionesJuradas:
          type: array
          items:
            $ref: "#/components/schemas/DeclaracionJuradaDto"
          description: Listado de declaraciones juradas
        descripcion:
          type: string
          description: Descripcion del tramite
        descripcionTramite:
          type: string
          description: Descripcion del procedimiento del tramite
        reglaPagoExencionAplicada:
          type: string
          description: Regla de pago o exención aplicada al tramite
        indNoRequierePago:
          type: boolean
          description: Indicador de si el tramite no requiere pago
    DeclaracionJuradaDto:
      type: object
      properties:
        id:
          type: integer
          description: Identificador de la declaracion jurada
        numeroDeclaracionJurada:
          type: string
          description: Numero de la declaracion jurada
        estadoDeclaracionJurada:
          type: string
          description: Identificador del estado de la declaracion jurada
        motivo:
          type: string
          description: Motivo de la declaracion jurada
        error:
          type: string
          description: Error de la declaracion jurada
        documentoNombre:
          type: string
          description: Nombre del documento
        fechaDeclaracionJurada:
          type: string
          format: date-time
          description: Fecha de la declaracion jurada
        fechaDenegacionDdjj:
          type: string
          format: date-time
          description: Fecha de denegacion de la declaracion jurada
        fechaAprobacionDdjj:
          type: string
          format: date-time
          description: Fecha de aprobacion de la declaracion jurada
...